Refactoring to optimise solution
